added component and auth files, added login/register form, split code
